requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
amqp==2.5.2
appdirs==1.4.3
apturl==0.5.2
asciinema==2.0.2
asn1crypto==0.24.0
astroid==2.3.0
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
autopep8==1.4.4
beautifulsoup4==4.9.0
billiard==3.6.3.0
boto3==1.10.32
botocore==1.13.50
Brlapi==0.6.6
bs4==0.0.1
cchardet==2.1.6
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coala-utils==0.5.1
colorama==0.4.3
command-not-found==0.3
configparser==5.0.0
constantly==15.1.0
crayons==0.3.0
cryptography==2.9.2
cssselect==1.1.0
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
elasticsearch==7.6.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
geographiclib==1.50
geopy==1.22.0
googletransx==2.4.2
greenlet==0.4.12
html2markdown==0.1.6.post0
html2text==2018.1.9
html5lib==0.999999999
httplib2==0.11.3
hyperlink==19.0.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.1.1
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
maxminddb==1.3.0
mccabe==0.6.1
msgpack==0.5.6
multidict==4.7.5
neovim==0.2.0
netifaces==0.10.4
numpy==1.18.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
openpyxl==2.5.12
pandas==1.0.3
parse==1.11.1
parsel==1.5.2
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.6.0
Protego==0.1.16
protobuf==3.0.0
proxybroker==0.3.1
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycares==3.1.1
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pyee==5.0.0
pygobject==3.26.1
PyHamcrest==2.0.2
pylint==2.4.1
pymacaroons==0.13.0
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyppeteer==0.0.25
pyquery==1.4.0
pyRFC3339==1.0
pyrsistent==0.16.0
PySocks==1.7.1
pytest==3.3.2
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
redis==3.1.0
reportlab==3.4.0
requests==2.23.0
requests-html==0.9.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
s3transfer==0.2.1
schedule==0.6.0
scour==0.36
Scrapy==2.0.1
scrapy-fake-useragent==1.2.0
scrapy-splash==0.7.2
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
simplejson==3.13.2
six==1.14.0
soupsieve==2.0
system-service==0.3
systemd-python==234
tabulate==0.8.5
thespian==3.9.3
tqdm==4.30.0
tweepy==3.8.0
twint==2.1.20
Twisted==20.3.0
twitterscraper==1.4.0
typed-ast==1.2.0
typing==3.7.4.1
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
vine==1.3.0
virtualenv==20.0.18
w3lib==1.21.0
wadllib==1.3.2
webdriver-manager==2.4.0
webencodings==0.5
websockets==7.0
Werkzeug==1.0.1
wrapt==1.11.2
xkit==0.0.0
yapf==0.28.0
yarl==1.4.2
you-get==0.4.1205
zipp==3.1.0
zope.interface==5.1.0
